+++ EBA/CLI/Config/BitcoinMapSpendsOptions.cs
@@ -1,0 +1,6 @@
+﻿namespace EBA.CLI.Config;
+
+public class BitcoinMapSpendsOptions
+{
+    public string BatchesFilename { init; get; }
+}
