+++ EBA/Graph/Bitcoin/Strategies/BlockNodeStrategy.cs
@@ -86,16 +86,6 @@ public class BlockNodeStrategy(bool serializeCompressed)
         return _mappings.GetCsv(node);
     }
 
-    public static BlockNode Deserialize(string[] values)
-    {
-        var properties = new Dictionary<string, object>();
-
-        for (var i = 0; i < _mappings.Length; i++)
-            properties[_mappings[i].Property.Name] = values[i];
-
-        return Deserialize(props: properties);
-    }
-
     // TODO: need a deserializer from string that returns only a given property,
     // so avoids building the whole object when only a few properties are needed
     // (e.g., for filtering in queries).
